Add convenience script when load testing
- `perf-check-status.sh`
Desgined to be run with `watch` or a for loop. Returns 1 if any output is not good, meaning it can be called in another automation script in the future.
- `perf-cleanup.sh`
Removes all inputs and outputs whose name starts with `Perftest`. This could be done with tags in the future, but this script is very useful as of now